Move utility components to ui subfolder
