+++ apps/client-react/src/components/ui/error-boundary.tsx
